tasking and messages refactored by each component with each store and actions - tested on Redux DevTools
