+++ redux-demo/src/app/messaging/messages/messages.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
-import { NgRedux, select } from 'ng2-redux'; 
-import { IAppState } from '../../store'; 
-import { INCREMENT, DECREMENT } from '../../actions'; 
+import { NgRedux, select } from '@angular-redux/store';
+import { IAppState } from '../../store';
+import { INCREMENT, DECREMENT } from '../../actions';
 
 @Component({
   selector: 'app-messages',
@@ -9,16 +9,16 @@ import { INCREMENT, DECREMENT } from '../../actions';
   styleUrls: ['./messages.component.css']
 })
 export class MessagesComponent  {
-  @select() newMessages; 
+  @select() newMessages;
 
   constructor(private ngRedux: NgRedux<IAppState>) {
   }
 
-  increment() { 
+  increment() {
     this.ngRedux.dispatch({ type: INCREMENT });
   }
 
-  decrement() { 
+  decrement() {
     this.ngRedux.dispatch({ type: DECREMENT });
   }
 }
